requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-demultiplexer==2.0.1
channels-redis==3.2.0
channelsmultiplexer==0.0.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-appconf==1.0.4
django-cors-headers==3.5.0
django-rest-firebase-auth==0.2.0
djangochannelsrestframework==0.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
docutils==0.16
firebase-admin==4.5.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
incremental==17.5.0
install==1.3.4
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.1
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pylama==7.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
rsa==4.6
rstcheck==3.3.1
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.0
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django==0.5.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
yapf==0.30.0
zope.interface==5.2.0
